v2 - Python/requirements.txt
# External deps - Debian:
# sudo apt-get build-dep rsync python3-bsddb3

gitpython
berkeleydb

# Non-deps:
# bsddb3
